refactor : 승패 결정 메소드 WinStatus로 이동
