+++ src/test/java/domain/blackjack/BlackJackTest.java
@@ -1,10 +1,6 @@
 package domain.blackjack;
 
-import domain.card.Card;
-import domain.card.Rank;
-import domain.card.Shape;
 import domain.participant.Dealer;
-import domain.participant.Participant;
 import domain.participant.Participants;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -29,41 +25,5 @@ class BlackJackTest {
                 () -> assertThat(participants.getValue().get(0).getCardCount()).isEqualTo(2),
                 () -> assertThat(participants.getValue().get(1).getCardCount()).isEqualTo(2)
         );
-    }
-
-    @DisplayName("참가자와 딜러 중 점수가 높은 사람이 승리한다.")
-    @Test
-    void isWinner() {
-        Dealer dealer = new Dealer();
-        Participants participants = new Participants(List.of("one", "two"));
-        BlackJack blackJack = new BlackJack(dealer, participants);
-
-        Participant participant = participants.getValue().get(0);
-        participant.receiveCard(new Card(Shape.HEART, Rank.KING));
-        participant.receiveCard(new Card(Shape.HEART, Rank.ACE));
-
-        dealer.receiveCard(new Card(Shape.HEART, Rank.QUEEN));
-
-        WinStatus winStatus = blackJack.isWinner(participant);
-        assertThat(winStatus).isEqualTo(WinStatus.WIN);
-    }
-
-    @DisplayName("참가자와 딜러의 점수가 같은 경우 카드의 장수가 더 적은 사람이 승리한다.")
-    @Test
-    void isWinnerWhenSameScore() {
-        Dealer dealer = new Dealer();
-        Participants participants = new Participants(List.of("one", "two"));
-        BlackJack blackJack = new BlackJack(dealer, participants);
-
-        Participant participant = participants.getValue().get(0);
-        participant.receiveCard(new Card(Shape.HEART, Rank.KING));
-        participant.receiveCard(new Card(Shape.HEART, Rank.THREE));
-        participant.receiveCard(new Card(Shape.HEART, Rank.EIGHT));
-
-        dealer.receiveCard(new Card(Shape.HEART, Rank.QUEEN));
-        dealer.receiveCard(new Card(Shape.HEART, Rank.ACE));
-
-        WinStatus winStatus = blackJack.isWinner(participant);
-        assertThat(winStatus).isEqualTo(WinStatus.LOSE);
     }
 }
